Simplify implementation, but reduce functionality
